requirements.txt
pip==22.0.4
numpy==1.25.2
pandas==2.1.0
gcloud==0.18.3
oauth2client==4.1.3
httplib2==0.15.0
google-api-python-client==1.6
google-cloud-bigquery==2.34.4
pyarrow==13.0.0
Werkzeug==2.2.2